購物車完成刪除系統檔產品查詢產品新增

+++ prjMvcDemo/Controllers/ProductController.cs
@@ -11,9 +11,14 @@ namespace prjMvcDemo.Controllers
         // GET: Product
         public ActionResult List()
         {
-            dbDemoEntities2 db = new dbDemoEntities2();
-            var datas = from c in db.tProduct
+            string keyword = Request.Form["txtKeyword"];
+            dbDemoEntities db = new dbDemoEntities();
+            IEnumerable<tProduct> datas = null;
+            if (string.IsNullOrEmpty(keyword))
+                datas = from c in db.tProduct
                         select c;
+            else
+                datas = db.tProduct.Where(p=>p.fName.Contains(keyword) || p.fQty.Value.ToString().Contains(keyword) );
             return View(datas);
         }
 
@@ -24,7 +29,7 @@ namespace prjMvcDemo.Controllers
         [HttpPost]
         public ActionResult Create( tProduct p)
         {
-            dbDemoEntities2 db = new dbDemoEntities2();
+            dbDemoEntities db = new dbDemoEntities();
             db.tProduct.Add(p);
             db.SaveChanges();
             return RedirectToAction("List");
@@ -32,11 +37,40 @@ namespace prjMvcDemo.Controllers
 
         public ActionResult Delete(int? id)
         {
-            dbDemoEntities2 db = new dbDemoEntities2();
+            if (id == null)
+                return RedirectToAction("List");
+            dbDemoEntities db = new dbDemoEntities();
             tProduct prod = db.tProduct.FirstOrDefault(t => t.fId == id);
             if (prod != null)
             {
                 db.tProduct.Remove(prod);
+                db.SaveChanges();
+            }
+            return RedirectToAction("List");
+        }
+
+        public ActionResult Edit(int? id)
+        {
+            if (id == null)
+                return RedirectToAction("List");
+           dbDemoEntities db = new dbDemoEntities();
+            tProduct prod = db.tProduct.FirstOrDefault(t => t.fId == id);
+            if (prod == null)
+                return RedirectToAction("List");
+            return View(prod);
+        }
+        [HttpPost]
+
+        public ActionResult Edit(tProduct p)
+        {
+            dbDemoEntities db = new dbDemoEntities();
+            tProduct prod = db.tProduct.FirstOrDefault(t => t.fId == p.fId);
+            if (prod != null)
+            {
+                prod.fName = p.fName;
+                prod.fQty = p.fQty;
+                prod.fCost = p.fCost;
+                prod.fPrice = p.fPrice;
                 db.SaveChanges();
             }
             return RedirectToAction("List");
